RavenDB-11849 - fix the cores calculation
